separate test set for CSV customization
