+++ src/test/resources/application-test.properties
@@ -1,7 +1,3 @@
-#supports "ru" and "en"
 cyclone.otusspring.language=en
-#file name will be constructed of {pollfile.base}_{language}.{pollfile.ext} properties
 cyclone.otusspring.pollfile.base=/test-poll
 cyclone.otusspring.pollfile.ext=csv
-cyclone.otusspring.pollfile.separator=|
-cyclone.otusspring.pollfile.comment=--
