+++ src/test/resources/custom-csv.properties
@@ -1,0 +1,5 @@
+cyclone.otusspring.language=en
+cyclone.otusspring.pollfile.base=/custom
+cyclone.otusspring.pollfile.ext=csv
+cyclone.otusspring.pollfile.separator=|
+cyclone.otusspring.pollfile.comment=--
