+++ src/test/resources/test-poll_en.csv
@@ -1,7 +1,7 @@
---this should be skipped
-First question?|aa1|aa1|bb1|cc1|dd1
-Second question with \| separator?|bb2|aa2|bb2|cc2|dd2
---this should be skipped
-Third question after comment?|cc3|aa3|bb3|cc3|dd3
-\--Fourth starts with comment symbol?|dd4|aa4|bb4|cc4|dd4
-Fifth question with separator in answers?|c\|c5|aa5|bb5|c\|c5|dd5
+#this should be skipped
+First question?,aa1,aa1,bb1,cc1,dd1
+Second question with \, separator?,bb2,aa2,bb2,cc2,dd2
+#this should be skipped
+Third question after comment?,cc3,aa3,bb3,cc3,dd3
+\#Fourth starts with comment symbol?,dd4,aa4,bb4,cc4,dd4
+Fifth question with separator in answers?,c\,c5,aa5,bb5,c\,c5,dd5
